backend/requirements.txt
fastapi
pydantic
starlette
python-dotenv
supabase
fastapi-limiter
tiktoken
markdownify
beautifulsoup4
openai
tqdm
pre-commit
graphviz
requests
fuzzysearch
anthropic
python-multipart
pypdf
python-pptx
python-docx
uvicorn
pymupdf
pytube
pydub
yt-dlp
percache
replicate
reportlab
gdown
ipykernel
httpx
langchain-openai
langchain